+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -1,7 +1,6 @@
 package edu.kis.vh.nursery;
 
-import edu.kis.vh.nursery.IIntStack;
-import edu.kis.vh.nursery.list.IntLinkedList;
+import edu.kis.vh.nursery.ownStackImplementations.IntLinkedList;
 
 public class DefaultCountingOutRhymer{
     private IIntStack intArrayStack;
